+++ wereldwijnen/src/main/java/be/vdab/repositories/SoortenRepository.java
@@ -7,12 +7,12 @@ import be.vdab.entities.Soorten;
 
 public class SoortenRepository extends AbstractRepository {
 	
-	public List<Soorten> findAllOfLand(int land) {
+	public List<Soorten> findAllOfLand(long land) {
 		return getEntityManager().createNamedQuery("Soorten.findAllOfLand", Soorten.class)
 				.setParameter("land", land).getResultList();
 	}
 	
-	public Soorten read(int id) {
+	public Soorten read(long id) {
 		return getEntityManager().find(Soorten.class, id);
 	}
