+++ wereldwijnen/src/main/java/be/vdab/services/SoortenService.java
@@ -8,11 +8,11 @@ import be.vdab.repositories.SoortenRepository;
 public class SoortenService extends AbstractService {
 	private final SoortenRepository soortenRepository = new SoortenRepository();
 	
-	public List<Soorten> findAllOfLand(int land){
+	public List<Soorten> findAllOfLand(long land){
 		return soortenRepository.findAllOfLand(land);
 	}
 	
-	public Soorten read(int id){
+	public Soorten read(long id){
 		return soortenRepository.read(id);
 	}
